+++ packages/integration-tests/api/api.js
@@ -232,30 +232,6 @@ async function getExecutionLogs({ prefix, executionName }) {
 }
 
 /**
- * Add a collection to Cumulus via the API
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {Object} params.collection - a collection object
- * @returns {Promise<Object>} - the POST confirmation from the API
- */
-async function addCollectionApi({ prefix, collection }) {
-  const response = await callCumulusApi({
-    prefix: prefix,
-    payload: {
-      httpMethod: 'POST',
-      resource: '/{proxy+}',
-      path: '/collections',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(collection)
-    }
-  });
-  return verifyCumulusApiResponse(response);
-}
-
-/**
  * Add a provider to Cumulus via the API
  *
  * @param {Object} params - params
@@ -465,7 +441,6 @@ module.exports = {
   getAsyncOperation,
   deletePdr,
   getExecutionLogs,
-  addCollectionApi,
   addProviderApi,
   getProviders,
   getCollections,